# Conflicts:
#	app/questions/repositories/question.py
#	app/questions/routes.py
#	app/questions/services/question.py

+++ app/questions/services/question.py
@@ -58,15 +58,32 @@ class QuestionService:
     async def update_question(
             self,
             question_id: int,
+            user_id: int,
             question_schema: QuestionUpdateSchema
-    ) -> QuestionOutSchema:
-        await self.question_repository.check_question_exists(question_id)
+    ) -> QuestionWithUserOutSchema:
+        question = await self.question_repository.get_question_if_exists(
+            question_id
+        )
+        if question.user_id != user_id:
+            raise HTTPException(
+                status_code=403,
+                detail='You are not allowed to update this question'
+            )
         await self.question_repository.update(question_id, question_schema)
-        return await self.question_repository.get_by_id_with_user(question_id)
+        await self.question_repository.expire_session_for_all()
+        return await self.question_repository.get_question_by_id(question_id)
 
     async def delete_question(
             self,
-            question_id: int
+            question_id: int,
+            user_id: int
     ) -> bool:
-        await self.question_repository.check_question_exists(question_id)
+        question = await self.question_repository.get_question_if_exists(
+            question_id
+        )
+        if question.user_id != user_id:
+            raise HTTPException(
+                status_code=403,
+                detail='You are not allowed to delete this question'
+            )
         return await self.question_repository.delete(question_id)
